Mensajes y correcto manejo de modales
