+++ app/api/estadisticos/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+estg_bp = Blueprint("est_graf", __name__)
+from . import routes
