+++ app/static/css/Coordinador/styleCoordinador.css
@@ -76,7 +76,7 @@ body {
 .sidebar {
   width: 270px;
   padding: 2rem;
-  border-right: 1px solid #0077c2;
+  border-right: 1px solid #0077c2;  
 }
 
 .sidebar h2,
@@ -245,30 +245,30 @@ button:hover,
 .modal {
   display: none;
   position: fixed;
-  top: 20%;
-  left: 50%;
-  transform: translate(-50%, -20%);
-  background-color: white;
-  border: 2px solid #0077c2;
-  padding: 20px;
-  z-index: 999;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  inset: 0;
+  background: rgba(0,0,0,0.6);
+  z-index: 1000;
+  align-items: center;
+  justify-content: center;
 }
 
 .modal.activo {
-  display: block;
+  display: flex;
 }
 
 .modal-content {
-  background-color: white;
-  padding: 20px;
-  width: 400px;
-  max-width: 90%;
-  border-radius: 8px;
   position: relative;
+  background: #fff;
+  padding: 30px;
+  width: 600px;      /* 🔑 Aquí controlas el ancho */
+  max-width: 90%;    /* 🔑 Responsivo */
+  max-height: 90vh;  /* 🔑 Para que no se desborde */
+  border-radius: 10px;
   border: 1px solid #0077c2;
-    overflow-y: auto; /* Scroll vertical si el contenido excede */
-
+  box-shadow: 0 4px 15px rgba(0,0,0,0.3);
+  overflow-y: auto;
+  animation: fadeIn 0.3s ease-in-out;
+  box-sizing: border-box;
 }
 
 .modal-content label {
@@ -282,8 +282,10 @@ button:hover,
   width: 100%;
   padding: 8px;
   margin-top: 4px;
+  margin-right: 20px ;
   border: 1px solid #ccc;
   border-radius: 4px;
+  box-sizing: border-box;
 }
 
 .close {
@@ -293,7 +295,13 @@ button:hover,
   font-size: 20px;
   cursor: pointer;
 }
-
+.close:hover {
+  color: red;
+}
+@keyframes fadeIn {
+  from { transform: scale(0.9); opacity: 0; }
+  to   { transform: scale(1); opacity: 1; }
+}
 
 /* ======================= */
 /* RESPONSIVE MOBILE STYLES */
@@ -466,22 +474,6 @@ button:hover,
   transform: translateY(0);
 }
 
-
-
-
-
-/* ...existing code... */
-.modal-content {
-  max-width: 95vw;
-  width: 400px;
-  max-height: 90vh;
-  overflow-y: auto;
-  margin: auto;
-  border-radius: 8px;
-  background: #fff;
-  padding: 20px;
-  box-sizing: border-box;
-}
 
 @media (max-width: 500px) {
   .modal-content {
